# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Yacht Dealers API",
        "description": "REST API for catamaran charters, bookings, guests, summaries, and personal requests in Dar es Salaam and Zanzibar.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API v1"
        }
    ],
    "paths": {
...
        "/catamarans/{catamaranId}/photos": {
            "post": {
                "tags": [
                    "Photos"
                ],
                "summary": "Add a photo to a catamaran",
                "operationId": "466e57df44f324c325461b7e6e18d662",
                "parameters": [
                    {
                        "name": "catamaranId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "path"
                                ],
                                "properties": {
                                    "path": {
                                        "type": "string",
                                        "example": "/storage/catamarans/ocean-spirit.jpg"
                                    },
                                    "caption": {
                                        "type": "string",
                                        "example": "Deck view"
                                    },
                                    "sort_order": {
                                        "type": "integer",
                                        "example": 0
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Photo created"
                    },
                    "404": {
                        "description": "Catamaran not found"
                    }
                }
            }
        },
        "/photos/{id}": {
            "delete": {
                "tags": [
                    "Photos"
                ],
                "summary": "Delete a catamaran photo",
                "operationId": "b156075f3d7445b32d9346f522ad6fc4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Photo deleted"
                    },
                    "404": {
                        "description": "Photo not found"
                    }
                }
            }
        },
...
        "/bookings/{bookingId}/guest": {
            "get": {
                "tags": [
                    "Guests"
                ],
                "summary": "Get guest details for a booking",
                "operationId": "c5c675da74ffaf8b36abf24f531f05fa",
                "parameters": [
                    {
                        "name": "bookingId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response"
                    },
                    "404": {
                        "description": "Booking or guest not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Guests"
                ],
                "summary": "Add or replace guest details for a booking",
                "operationId": "37850c4113ba8b2f44b8f5cd55737219",
                "parameters": [
                    {
                        "name": "bookingId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "date",
                                    "phone_number",
                                    "email",
                                    "number_of_passengers"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Utwah Mwingira"
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-06-17"
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "example": "0613062067"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "guest@example.com"
                                    },
                                    "number_of_passengers": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Guest saved"
                    },
                    "404": {
                        "description": "Booking not found"
                    }
                }
            }
        },
        "/guests/{id}": {
            "put": {
                "tags": [
                    "Guests"
                ],
                "summary": "Update guest details",
                "operationId": "978244f7244a2b29af0c3517d31af9ac",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "phone_number": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "number_of_passengers": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Guest updated"
                    },
                    "404": {
                        "description": "Guest not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Guests"
                ],
                "summary": "Delete guest details",
                "operationId": "33f7f1b8ea19ac0915228fa43a275ec9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Guest deleted"
                    },
                    "404": {
                        "description": "Guest not found"
                    }
                }
            }
        },
...
    },
...
}